+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * Added `content_type` and `content_encoding` configuration for
    the Matcher plugin - #37
+* Added header options to set message headers based on the message content on
+  the Matcher plugin - #35
 
 # 0.10.x